requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
python-multipart==0.0.17
pydantic==2.10.3
python-dotenv==1.0.1